% Generated by roxygen2 (4.0.1): do not edit by hand
\name{onemode.projection}
\alias{onemode.projection}
\title{One-mode (actor) projection of a two-mode (affiliation) network}
\usage{
onemode.projection(bigraph, type = 0, name = "name")
}
\arguments{
\item{bigraph}{The ambient affiliation network}

\item{type}{The actor node type in bigraph}

\item{name}{What attribute of the actor nodes in bigraph to name the nodes
in the projection (defaults to 'name')}
}
\description{
This function uses the igraph function bipartite.projection to compute the
projection of a two-mode network onto the desired node type.
}

