% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wedge.dependency.hist}
\alias{wedge.dependency.hist}
\title{Wedge-dependency histograms}
\usage{
wedge.dependency.hist(wedges, max.no = 12, wedge.nos = NULL)
}
\arguments{
\item{wedges}{A wedge list of two columns: wedge count and closed wedge count}

\item{max.no}{An upper bound on the wedge count to construct histograms for}

\item{wedge.nos}{Specific wedge counts to include (overrides max.no)}
}
\description{
Draw cool strata of histograms of wedge closure rate by wedge count.
}

